requirements.txt
requests==2.22.0
Flask==1.0.2
python-dotenv==0.10.3
